+++ docs/CHANGE_LOG.md
@@ -9,7 +9,8 @@
 - Added a new binding/option called "pauseTimeoutOnChange", which states the milliseconds to wait for new changes to be detected when a change is detected (defaults to 0, which is off).
 
 #### **Public Functions:**
-- Added new public function "cancelAllWatches()", which is used to cancel all the watches currently running, or paused.
+- Added new public function "cancelWatches()", which is used to cancel all the watches currently running, or paused.
+- Added new public function "getWatches()", which is used to return all the watches currently running, or paused.
 
 #### **Documentation:**
 - Fixed some of the documentation spelling/grammar mistakes.